+++ lib/jxl/decode.cc
@@ -2148,7 +2148,6 @@ JxlDecoderStatus JxlDecoderGetBasicInfo(const JxlDecoder* dec,
 
     info->have_preview = meta.have_preview;
     info->have_animation = meta.have_animation;
-    info->have_intrinsic_size = meta.have_intrinsic_size;
     info->orientation = static_cast<JxlOrientation>(meta.orientation);
 
     if (!dec->keep_orientation) {
@@ -2192,9 +2191,12 @@ JxlDecoderStatus JxlDecoderGetBasicInfo(const JxlDecoder* dec,
       info->animation.have_timecodes = dec->metadata.m.animation.have_timecodes;
     }
 
-    if (info->have_intrinsic_size) {
-      info->intrinsic_size.xsize = dec->metadata.m.intrinsic_size.xsize();
-      info->intrinsic_size.ysize = dec->metadata.m.intrinsic_size.ysize();
+    if (meta.have_intrinsic_size) {
+      info->intrinsic_xsize = dec->metadata.m.intrinsic_size.xsize();
+      info->intrinsic_ysize = dec->metadata.m.intrinsic_size.ysize();
+    } else {
+      info->intrinsic_xsize = info->xsize;
+      info->intrinsic_ysize = info->ysize;
     }
   }